pyproject.toml
[tool.poetry]
name = "openapi-rdf-converter"
version = "0.1.0"
description = "Convert OpenAPI YAML specifications to RDF vocabularies and SHACL validation shapes"
authors = ["Jean Martins <jean.martins@ericsson.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.6"
pyyaml = "^6.0.2"
rdflib = "^7.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
